requirements/prod.txt
Django==1.5.1
Fabric==1.6.0
MySQL-python==1.2.4
South==0.7.6
distribute==0.6.36
django-autoslug==1.6.1
django-model-utils==1.3.1
django-tastypie==0.9.14
mimeparse==0.1.3
paramiko==1.10.1
pycrypto==2.6
python-dateutil==2.1
six==1.3.0
stevedore==0.8
wsgiref==0.1.2
